Refactor client codes in webhook and recovery backend services
Longhorn 2293
